requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
pytest>=8.0.0
pytest-mock>=3.12.0
responses>=0.25.0
